Moved logout to global functions
