+++ src/lib/authentication/authValidator.ts
@@ -4,27 +4,16 @@ import type { LoginTokens } from '$lib/authentication/loginStore';
 import { get } from "svelte/store";
 import reauthenticate from "../-universis/authenticator-deprecated/reauthenticate.js";
 import { Network } from '@capacitor/network';
-import { Preferences } from "@capacitor/preferences";
-import Dexie from "dexie";
 import OIDCClient from "./OIDCClient.js";
 import { Capacitor } from "@capacitor/core";
 import Config from "$src/app.config";
-import {goto} from "$app/navigation";
-import { Browser } from '@capacitor/browser'; 
-import { resetPersistedStores } from "$src/routes/persistedStoreDeclarations.js";
+
+
 
 const isMobile = Capacitor.isNativePlatform();
 const authClient = new OIDCClient(Config.auth);
 
-// Do we wanna log out? Let's clear our path
-export async function invalidateAuth(){
-    let x = "https://oauth2.it.auth.gr/auth/realms/universis/protocol/openid-connect/logout" //?id_token_hint=" + get(loginStore).id_token;
-    await resetPersistedStores();
-    localStorage.clear();
-    Preferences.clear();
-    Dexie.delete('cachedData');
-    await Browser.open({ url: x });
-}
+
 
 export async function judgeAuth() {
     /*
